apps/covibe.ai/backend/requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Redis & Caching
redis==5.0.1
celery==5.3.6
hiredis==2.3.2

# AI & ML
openai==1.10.0
anthropic==0.8.1
google-generativeai==0.3.2
langchain==0.1.4
tiktoken==0.5.2

# Code Analysis
tree-sitter==0.20.4
tree-sitter-python==0.20.4
tree-sitter-javascript==0.20.3
tree-sitter-typescript==0.20.5
pylint==3.0.3
black==24.1.1
isort==5.13.2

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# HTTP & WebSocket
httpx==0.26.0
websockets==12.0
aiohttp==3.9.1

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
python-dateutil==2.8.2
pytz==2024.1

# Monitoring & Logging
sentry-sdk==1.40.0
loguru==0.7.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Development
ipython==8.20.0
